+++ test/cases/phase4/closures/c-output/expected.c
@@ -11,6 +11,6 @@ int main(void) {
     const int32_t factor = 3;
     _closure_0_env mul_env = {.factor = factor};
     tsc_closure mul = {.env = &mul_env, .fn = (void*)_closure_0_fn};
-    printf("%d\n", ((int32_t (*)(void *, double))mul.fn)(mul.env, 7));
+    printf("%d\n", ((int32_t (*)(void *, int32_t))mul.fn)(mul.env, 7));
     return 0;
 }
